[#43]Refactor(apis): pagination 타입 comments -> apis 타입 이용하도록 수정. (타입 내용은 같음.)
