--- types/commons.ts
@@ -1,8 +1,0 @@
-export interface Pagination {
-  currentPage: number;
-  totalPages: number;
-  totalCount: number;
-  pageSize: number;
-  hasNext: boolean;
-  hasPrevious: boolean;
-}
